perf: Remove unecessary wrangling code
This was an artifact of the cookies work and concluded that, primarily, the need to remove 20th December and use bespoke 'monthly' periods was too confusing for the end user. Thus, this code has been greatly simplified. Also have done things like remove ignoring NAs and unecessary coding of pageviews column as numeric.

This follows a code review.