Repeat non-corona page detection for 24 March
Redo this analysis with data from 24 March, the first day following the Prime
Minister's 'lockdown' message.

I've replaced the graph and table with a new graph that shows the
covid-relatedness ratio vs the total hits per page.  This is easier to explore,
and shows that pages related to covid sessions are generally not viewed much.