Update pre-commit
This commit updates `pre-commit` to use the `detect-secrets` module to detect anything which looks like a secret.

It also updates the instructions for installing `pre-commit` so that it is now run on a per-commit basis (not per-push). This is to ensure better sanitisation of commits (to avoid edge cases where sensitive data could otherwise leak through).