Update Sphinx documentation with `src.make_feedback_tool_data` functions
Added Markdown support for Sphinx, and also added API references to the
functions as well. Minor edit it `create_phrase_level_columns`
docstring to resolve Sphinx formatting issues.

Also include a `docs/README.md` to explain how to write documentation
for Sphinx, and a section in `README.md` on how to build and view this
documentation.